refactor: split landing.rs, kobo.rs, and settings/mod.rs under the 800-line cap (#1642)
Three production files had crossed the ~800-line file-shape cap. Each is
split by sub-topic into a sibling-module directory, mirroring the existing
books/ and metadata_lookup/ precedents — pure code motion, no behavior
change:

- db/src/settings/mod.rs (811 lines): provider-key KV helpers
  (get/set/effective/status/seed for the Hardcover and Google Books keys,
  plus provider_keys()) move to a new settings/keys.rs, re-exported from
  mod.rs. 811 -> 693 lines.

- server/src/backend/kobo.rs (846 lines) becomes kobo/mod.rs (router
  assembly + shared helpers only), with the protocol phases split into
  kobo/auth.rs (handshake/token exchange), kobo/sync.rs (library sync +
  read-state GET), kobo/state.rs (read-state PUT), and kobo/resources.rs
  (download/cover/tags) — consistent with the existing kobo/reading_services/
  and kobo/store_resources.rs submodules. Router-building function
  signatures are unchanged.

- frontend/src/pages/landing.rs (958 lines): split into landing/signals.rs
  (reactive state + fetch/effect wiring), landing/view.rs (per-render
  derived state + event handlers), and landing/body.rs (mobile/web
  presentation branches + bulk-edit helpers), leaving landing.rs as a thin
  composition root. SSR/WASM markup is unchanged (pure code motion).
